backend/requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python-headless
tensorflow
python-multipart